pyproject.toml
[project]
name = "web-crawler-ut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.68.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "asyncio>=3.4.3",
    "pathlib>=1.0.1"
]
